Refactor imports in afip_service and pdf_generator to use BASE_DIR and STATIC_DIR from server module
