+++ server/core/services/pdf_generator.py
@@ -12,8 +12,9 @@ from reportlab.lib import colors
 from reportlab.lib.utils import ImageReader
 from num2words import num2words
 
+from server import STATIC_DIR
 from server.core.models import Venta, EstadoVenta
-from server.config import STATIC_DIR
+
 
 locale.setlocale(locale.LC_ALL, "es_AR.UTF-8")
